REF: move json request stuff into db_backed
